refactor: helper 폴더 내의 폴더 명 지정 및 위치 조정
IntersectionObserver를 helper내로 옮기고 helper 폴더 하위에 많아진 파일 개수에 따른 폴더로 묶는 것 진행
