+++ frontend/src/App.tsx
@@ -4,8 +4,8 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 
 import Loading from '@/components/@common/Loading/Loading';
 import SnackBar from '@/components/@common/SnackBar/SnackBar';
-import PrivateRouter from '@/components/@helper/PrivateRouter';
-import PublicRouter from '@/components/@helper/PublicRouter';
+import PrivateRouter from '@/components/@helper/router/PrivateRouter';
+import PublicRouter from '@/components/@helper/router/PublicRouter';
 import Header from '@/components/@layout/Header/Header';
 import TabBar from '@/components/@layout/TabBar/TabBar';
 import { URL } from '@/constants/url';
