+++ frontend/src/components/@helper/errorBoundary/CustomError.tsx
