+++ frontend/src/components/@helper/errorBoundary/LogicErrorBoundary.tsx
@@ -1,7 +1,7 @@
 import { PropsWithStrictChildren } from 'gongseek-types';
 
-import CommonErrorBoundary from '@/components/@helper/CommonErrorBoundary';
-import CustomError from '@/components/@helper/CustomError';
+import CommonErrorBoundary from '@/components/@helper/errorBoundary/CommonErrorBoundary';
+import CustomError from '@/components/@helper/errorBoundary/CustomError';
 import { ACCESSTOKEN_KEY } from '@/constants';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import { URL } from '@/constants/url';
