+++ frontend/src/components/@helper/observer/InfiniteScrollObserver.tsx
